Support config cache
